Enforce search behavior according to &foldopen
Sneaking now behaves like / when "foldopen-=search" is set: Closed folds
are considered to be single targets and are not opened when sneaking to
them. They are opened, however, when the default-&foldopen is used or
"foldopen+=search" is set.
Resolve #102.

This does not affect streak-mode (which still ignores &foldopen).