Simplify highlighting: remove SneakCursor
`SneakCursor` was added as a workaround for the case where the user's
`Cursor` highlight was broken. But with 'termguicolors', there are more
cases to consider (also: `has('gui_running')` is meaningless in Nvim).

Instead of adding more handling, just use `Cursor`. If the user has
a broken Cursor highlight, it's better that they fix it.