dot-repeat of 'operatorfunc' (g@)
Use an `OptionSet` handler to remember the current 'operatorfunc' (g@)
operation. If the same operation is seen later, repeat the last input
instead of prompting for input. This works because 'operatorfunc'
(apparently) is _not_ set on dot-repeat.

This is needed because our `repeat#set()` (vim-repeat plugin) gets
clobbered in some cases (such as vim-surround) where the motion (Sneak)
happens in the middle of the operation.

fixes #94